moved GET tweets into tweet_routes.py
